Removed one-thread-per-drive
Now there is only a single thread that runs a loop with a frequency of
DriveManager.cpp/RESOLUTION Hertz. This thread will take care of every
drive.

The reason for this is that it seemed like the RPi, while handling ~3
drives just well, couldn't handle ~6 drives, so it sounded bad (at least
here). The system now can only play frequencies that can be expressed as
a fraction of the resolution frequency, i.e. that satisfy
    res = freq * n (n is a natural number)

This means that some notes may sound off, because the accuracy that we
had with every single thread is gone. As a tradeoff we can now use more
drives and let it sound even more awesome!